refactor: get rid of package expressions by moving the code into package expression
