+++ lib/checker/expression/http_headers.go
@@ -1,4 +1,4 @@
-package expressions
+package expression
 
 import (
 	"net/http"
