+++ lib/checker/expression/http_headers_test.go
@@ -1,4 +1,4 @@
-package expressions
+package expression
 
 import (
 	"net/http"
