+++ lib/checker/expression/loadavg.go
@@ -1,4 +1,4 @@
-package expressions
+package expression
 
 import (
 	"context"
