+++ lib/checker/expression/url_values.go
@@ -1,4 +1,4 @@
-package expressions
+package expression
 
 import (
 	"errors"
