+++ lib/checker/expression/url_values_test.go
@@ -1,4 +1,4 @@
-package expressions
+package expression
 
 import (
 	"net/url"
